O3-4771:Additional scenarios
